+++ docs/changelog.rst
@@ -21,8 +21,8 @@ and ``.flex`` files now return a collection.
 - Adds ``parse_cdl`` which returns a :class:`ColorCollection` .
 - Adds ``write_cdl`` which returns a :class:`ColorCollection` as a ``cdl`` file.
 - :class:`ColorCollection` is now a fully functional container class, with many attributes and methods.
-- Added :class:`ColorDecision` , which stores either a :class:`ColorCorrection` or :class:`ColorCorrectionReference` , and an optional :class:`MediaRef`
-- Added :class:`ColorCorrectionReference` , which stores a reference to a :class:`ColorCorrection`
+- Added :class:`ColorDecision` , which stores either a :class:`ColorCorrection` or :class:`ColorCorrectionRef` , and an optional :class:`MediaRef`
+- Added :class:`ColorCorrectionRef` , which stores a reference to a :class:`ColorCorrection`
 - Added ``parent`` attribute to :class:`ColorCorrection` .
 - Calling ``sop_node`` or ``sat_node`` on a :class:`ColorCorrection` before attempting to set a SOP or Sat power now works.
 - :class:`ColorCorrection` ``cdl_file`` init argument renamed to ``input_file``, which is now optional and able to be set after init.